move util from netsuite to lib
